move asterisks extraction to TranslateQualifiedNamesVisitor (before column names normalisation)
