+++ dbms/src/Interpreters/QueryNormalizer.h
@@ -1,11 +1,9 @@
 #pragma once
 
-#include <unordered_set>
 #include <map>
 
 #include <Core/Names.h>
 #include <Parsers/IAST.h>
-#include <Interpreters/DatabaseAndTableWithAlias.h>
 #include <Interpreters/Aliases.h>
 
 namespace DB
@@ -21,9 +19,9 @@ inline bool functionIsInOrGlobalInOperator(const String & name)
     return functionIsInOperator(name) || name == "globalIn" || name == "globalNotIn";
 }
 
+class ASTSelectQuery;
 class ASTFunction;
 class ASTIdentifier;
-class ASTExpressionList;
 struct ASTTablesInSelectQueryElement;
 class Context;
 
@@ -53,10 +51,6 @@ public:
 
         const Aliases & aliases;
         const ExtractedSettings settings;
-        const Context * context;
-        const NameSet * source_columns_set;
-        const std::vector<TableWithColumnNames> * tables_with_columns;
-        std::unordered_set<String> join_using_columns;
 
         /// tmp data
         size_t level;
@@ -64,26 +58,11 @@ public:
         SetOfASTs current_asts;     /// vertices in the current call stack of this method
         std::string current_alias;  /// the alias referencing to the ancestor of ast (the deepest ancestor with aliases)
 
-        Data(const Aliases & aliases_, ExtractedSettings && settings_, const Context & context_,
-             const NameSet & source_columns_set, const std::vector<TableWithColumnNames> & tables_with_columns_)
-            : aliases(aliases_)
-            , settings(settings_)
-            , context(&context_)
-            , source_columns_set(&source_columns_set)
-            , tables_with_columns(&tables_with_columns_)
-            , level(0)
-        {}
-
         Data(const Aliases & aliases_, ExtractedSettings && settings_)
             : aliases(aliases_)
             , settings(settings_)
-            , context(nullptr)
-            , source_columns_set(nullptr)
-            , tables_with_columns(nullptr)
             , level(0)
         {}
-
-        bool processAsterisks() const { return tables_with_columns && !tables_with_columns->empty(); }
     };
 
     QueryNormalizer(Data & data)
@@ -102,13 +81,10 @@ private:
 
     static void visit(ASTIdentifier &, ASTPtr &, Data &);
     static void visit(ASTFunction &, const ASTPtr &, Data &);
-    static void visit(ASTExpressionList &, const ASTPtr &, Data &);
     static void visit(ASTTablesInSelectQueryElement &, const ASTPtr &, Data &);
     static void visit(ASTSelectQuery &, const ASTPtr &, Data &);
 
     static void visitChildren(const ASTPtr &, Data & data);
-
-    static void extractJoinUsingColumns(const ASTPtr ast, Data & data);
 };
 
 }
