+++ dbms/src/Interpreters/TranslateQualifiedNamesVisitor.cpp
@@ -1,3 +1,5 @@
+#include <Poco/String.h>
+
 #include <Interpreters/TranslateQualifiedNamesVisitor.h>
 #include <Interpreters/IdentifierSemantic.h>
 
@@ -5,10 +7,14 @@
 #include <Core/Names.h>
 
 #include <Parsers/ASTIdentifier.h>
+#include <Parsers/ASTAsterisk.h>
 #include <Parsers/ASTQualifiedAsterisk.h>
 #include <Parsers/ASTSelectQuery.h>
 #include <Parsers/ASTSelectWithUnionQuery.h>
 #include <Parsers/ASTTablesInSelectQuery.h>
+#include <Parsers/ASTExpressionList.h>
+#include <Parsers/ASTLiteral.h>
+#include <Parsers/ASTFunction.h>
 
 
 namespace DB
@@ -41,12 +47,14 @@ std::vector<ASTPtr *> TranslateQualifiedNamesMatcher::visit(ASTPtr & ast, Data &
 {
     if (auto * t = typeid_cast<ASTIdentifier *>(ast.get()))
         return visit(*t, ast, data);
-    if (auto * t = typeid_cast<ASTQualifiedAsterisk *>(ast.get()))
-        return visit(*t, ast, data);
     if (auto * t = typeid_cast<ASTTableJoin *>(ast.get()))
         return visit(*t, ast, data);
     if (auto * t = typeid_cast<ASTSelectQuery *>(ast.get()))
         return visit(*t, ast, data);
+    if (auto * node = typeid_cast<ASTExpressionList *>(ast.get()))
+        visit(*node, ast, data);
+    if (auto * node = typeid_cast<ASTFunction *>(ast.get()))
+        visit(*node, ast, data);
     return {};
 }
 
@@ -71,6 +79,18 @@ std::vector<ASTPtr *> TranslateQualifiedNamesMatcher::visit(ASTIdentifier & iden
     }
 
     return {};
+}
+
+/// As special case, treat count(*) as count(), not as count(list of all columns).
+void TranslateQualifiedNamesMatcher::visit(ASTFunction & node, const ASTPtr &, Data &)
+{
+    ASTPtr & func_arguments = node.arguments;
+
+    String func_name_lowercase = Poco::toLower(node.name);
+    if (func_name_lowercase == "count" &&
+        func_arguments->children.size() == 1 &&
+        typeid_cast<const ASTAsterisk *>(func_arguments->children[0].get()))
+        func_arguments->children.clear();
 }
 
 std::vector<ASTPtr *> TranslateQualifiedNamesMatcher::visit(const ASTQualifiedAsterisk & , const ASTPtr & ast, Data & data)
@@ -100,8 +120,11 @@ std::vector<ASTPtr *> TranslateQualifiedNamesMatcher::visit(ASTTableJoin & join,
     return out;
 }
 
-std::vector<ASTPtr *> TranslateQualifiedNamesMatcher::visit(ASTSelectQuery & select, const ASTPtr & , Data &)
+std::vector<ASTPtr *> TranslateQualifiedNamesMatcher::visit(ASTSelectQuery & select, const ASTPtr & , Data & data)
 {
+    if (auto join = select.join())
+        extractJoinUsingColumns(join->table_join, data);
+
     /// If the WHERE clause or HAVING consists of a single qualified column, the reference must be translated not only in children,
     /// but also in where_expression and having_expression.
     std::vector<ASTPtr *> out;
@@ -112,6 +135,111 @@ std::vector<ASTPtr *> TranslateQualifiedNamesMatcher::visit(ASTSelectQuery & sel
     if (select.having_expression)
         out.push_back(&select.having_expression);
     return out;
+}
+
+/// Replace *, alias.*, database.table.* with a list of columns.
+void TranslateQualifiedNamesMatcher::visit(ASTExpressionList & node, const ASTPtr &, Data & data)
+{
+    const auto & tables_with_columns = data.tables;
+    const auto & source_columns = data.source_columns;
+
+    ASTs old_children;
+    if (data.processAsterisks())
+    {
+        bool has_asterisk = false;
+        for (const auto & child : node.children)
+        {
+            if (typeid_cast<const ASTAsterisk *>(child.get()))
+            {
+                if (tables_with_columns.empty())
+                    throw Exception("An asterisk cannot be replaced with empty columns.", ErrorCodes::LOGICAL_ERROR);
+                has_asterisk = true;
+                break;
+            }
+            else if (auto qa = typeid_cast<const ASTQualifiedAsterisk *>(child.get()))
+            {
+                visit(*qa, child, data); /// check if it's OK before rewrite
+                has_asterisk = true;
+                break;
+            }
+        }
+
+        if (has_asterisk)
+        {
+            old_children.swap(node.children);
+            node.children.reserve(old_children.size());
+        }
+    }
+
+    for (const auto & child : old_children)
+    {
+        if (typeid_cast<const ASTAsterisk *>(child.get()))
+        {
+            bool first_table = true;
+            for (const auto & [table_name, table_columns] : tables_with_columns)
+            {
+                for (const auto & column_name : table_columns)
+                    if (first_table || !data.join_using_columns.count(column_name))
+                    {
+                        /// qualifed names for duplicates
+                        if (!first_table && source_columns.count(column_name))
+                            node.children.emplace_back(std::make_shared<ASTIdentifier>(table_name.getQualifiedNamePrefix() + column_name));
+                        else
+                            node.children.emplace_back(std::make_shared<ASTIdentifier>(column_name));
+                    }
+
+                first_table = false;
+            }
+        }
+        else if (const auto * qualified_asterisk = typeid_cast<const ASTQualifiedAsterisk *>(child.get()))
+        {
+            DatabaseAndTableWithAlias ident_db_and_name(qualified_asterisk->children[0]);
+
+            bool first_table = true;
+            for (const auto & [table_name, table_columns] : tables_with_columns)
+            {
+                if (ident_db_and_name.satisfies(table_name, true))
+                {
+                    for (const auto & column_name : table_columns)
+                    {
+                        /// qualifed names for duplicates
+                        if (!first_table && source_columns.count(column_name))
+                            node.children.emplace_back(std::make_shared<ASTIdentifier>(table_name.getQualifiedNamePrefix() + column_name));
+                        else
+                            node.children.emplace_back(std::make_shared<ASTIdentifier>(column_name));
+                    }
+                    break;
+                }
+
+                first_table = false;
+            }
+        }
+        else
+            node.children.emplace_back(child);
+    }
+}
+
+/// 'select * from a join b using id' should result one 'id' column
+void TranslateQualifiedNamesMatcher::extractJoinUsingColumns(const ASTPtr ast, Data & data)
+{
+    const auto & table_join = typeid_cast<const ASTTableJoin &>(*ast);
+
+    if (table_join.using_expression_list)
+    {
+        auto & keys = typeid_cast<ASTExpressionList &>(*table_join.using_expression_list);
+        for (const auto & key : keys.children)
+            if (auto opt_column = getIdentifierName(key))
+                data.join_using_columns.insert(*opt_column);
+            else if (typeid_cast<const ASTLiteral *>(key.get()))
+                data.join_using_columns.insert(key->getColumnName());
+            else
+            {
+                String alias = key->tryGetAlias();
+                if (alias.empty())
+                    throw Exception("Logical error: expected identifier or alias, got: " + key->getID(), ErrorCodes::LOGICAL_ERROR);
+                data.join_using_columns.insert(alias);
+            }
+    }
 }
 
 }
