+++ dbms/src/Interpreters/TranslateQualifiedNamesVisitor.h
@@ -13,6 +13,8 @@ class ASTIdentifier;
 class ASTQualifiedAsterisk;
 struct ASTTableJoin;
 class ASTSelectQuery;
+class ASTExpressionList;
+class ASTFunction;
 
 /// Visit one node for names qualification. @sa InDepthNodeVisitor.
 class TranslateQualifiedNamesMatcher
@@ -22,15 +24,26 @@ public:
     {
         const NameSet & source_columns;
         const std::vector<TableWithColumnNames> & tables;
+        std::unordered_set<String> join_using_columns;
+        bool has_columns;
 
-        static void setTablesOnly(const std::vector<DatabaseAndTableWithAlias> & tables,
-                                  std::vector<TableWithColumnNames> & tables_with_columns)
+        Data(const NameSet & source_columns_, const std::vector<TableWithColumnNames> & tables_, bool has_columns_ = true)
+            : source_columns(source_columns_)
+            , tables(tables_)
+            , has_columns(has_columns_)
+        {}
+
+        static std::vector<TableWithColumnNames> tablesOnly(const std::vector<DatabaseAndTableWithAlias> & tables)
         {
-            tables_with_columns.clear();
+            std::vector<TableWithColumnNames> tables_with_columns;
             tables_with_columns.reserve(tables.size());
+
             for (const auto & table : tables)
                 tables_with_columns.emplace_back(TableWithColumnNames{table, {}});
+            return tables_with_columns;
         }
+
+        bool processAsterisks() const { return !tables.empty() && has_columns; }
     };
 
     static constexpr const char * label = "TranslateQualifiedNames";
@@ -43,10 +56,14 @@ private:
     static std::vector<ASTPtr *> visit(const ASTQualifiedAsterisk & node, const ASTPtr & ast, Data &);
     static std::vector<ASTPtr *> visit(ASTTableJoin & node, const ASTPtr & ast, Data &);
     static std::vector<ASTPtr *> visit(ASTSelectQuery & node, const ASTPtr & ast, Data &);
+    static void visit(ASTExpressionList &, const ASTPtr &, Data &);
+    static void visit(ASTFunction &, const ASTPtr &, Data &);
+
+    static void extractJoinUsingColumns(const ASTPtr ast, Data & data);
 };
 
 /// Visits AST for names qualification.
-/// It finds columns (general identifiers and asterisks) and translate their names according to tables' names.
+/// It finds columns and translate their names to the normal form. Expand asterisks and qualified asterisks with column names.
 using TranslateQualifiedNamesVisitor = InDepthNodeVisitor<TranslateQualifiedNamesMatcher, true>;
 
 }
